+++ searchbase/lib/searchbase.dart
@@ -1,4 +1,4 @@
 export 'src/searchbase.dart';
-export 'src/searchcomponent.dart';
+export 'src/searchwidget.dart';
 export 'src/constants.dart';
 export 'src/types.dart';