+++ searchbase/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.0-alpha
 
-- Add `SearchComponent` and `SearchBase` classes
+- Add `SearchWidget` and `SearchBase` classes
 - Implement query generation and analytics
 - Recent searches
 - Popular Suggestions